+++ docs/HISTORY.rst
@@ -4,7 +4,8 @@ Changelog
 0.1.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- cookieconsent cookie now expires in 10 years
+  [cekk]
 
 
 0.1.1 (2015-10-14)